BEIS users can add planned disbursements anytime
BEIS users should be able to add planned disbursements to an activity
when there is no editable report for the activity. Because the report
association is required they cannot.

Just like for transactions, make the association optional but validate
the presence of the value if the activity is for a delivery partner.

This change removes the shoulda spec for the report association as it is
now covered by the validations spec i.e. we check that the association
is set or not based on the activity organisation.