Upload transactions in bulk as CSV
This exposes the functionality of ImportTransactions via a file upload
that accepts CSV. The CSV is parsed in the controller and the resulting
rows fed into the importer, which itself is thoroughly unit-tested.

If any errors occur, then no transactions are imported, and the errors
are displayed to the user. If all the rows are valid, then the
transactions are saved and a success message is displayed.