Refactor Activity.projects_and_third_party_projects_for_report
The implementation of this method would be problematic if used for a
bulk operation involving lots of activities under the same report. It
loads all the level C/D Activity records for a report's Organisation,
and then further filters them by checking whether the Activity's
`associated_fund` matches that of the Report.

`Activity#associated_fund` requires 2 or 3 further queries to look up
the fund under which the level C/D activity sites. So if the initial
query returned 500 activities, this would add an additional 1,000-1,500
queries to load the fund for each result, which will often be the same
object.

Instead, we invert this as follows. Given the fund ID from the Report,
we can get the IDs of all programmes (level B) under this fund. From
these we can get the IDs of all projects (level C) under those
programmes.

From these two lists of IDs, we can load all the level C and D
activities that have one of them as a parent. The query ends up looking
like this?

    SELECT *
    FROM activities
    WHERE organisation_id = ?
    AND (level = 'project' AND parent_id IN (<programme_ids>))
    OR (level = 'third_party_project' AND parent_id IN (<project_ids>))

This reduces the lookup for all relevant activities to three queries.