`collaboration type` added to create activity
This is a new field required when users create new activities through RODA.
It is required for levels B, C, and D. It is optional for ingested activities.
The client requested to have this option as radio buttons with a pre-selected
value of `Bilateral`, since it is the most common value for this field on
the existing activities.

`collaboration_type` gets reported to IATI through the XMLs, so a later commit
will contain the logic necessary to make this happen.