Adding `collaboration-type` to activity XMLs
Now that `collaboration_type` has been added to the create activity form and
considering that this field does get reported to IATI, we have included this
in the XML, following IATI guidance.

Apart from the specs, I have manually checked several files on the new IATI
validator to make sure they are compliant with the requirements.