Pass Report explicitly from ImportTransactions to CreateTransaction
The plan is that the bulk transaction upload process will be initiated
from a report's page, and so we'll pass the report in question into the
ImportTransactions class so it knows which report to add transactions
to.

We do this by passing a Report into ImportTransactions, and forwarding
this on to CreateTransaction following the refactor in the previous
commit.

This is not just a correctness improvement (making sure all transactions
are added to the same report without needing to lock anything), it's a
performance improvement too. Without this, each transaction added via
CreateTransaction would trigger a new query to look the Report for its
Activity up. All of these should result in the same Report record, so
doing this for every row is wasteful.