IngestCsv creates activities when a parent is detected
When the CSV includes a column named "parent_roda_identifier_compound",
IngestCsv will create new Activity for each row, parented to the
Activity that matches it, and populated with default values.

We then can re-use any of the existing "process_*" methods to clean up
the data as usual, before saving these back into the database.