Improve logging when rows are skipped
It was difficult to read the logs for activities that were skipped
entirely because of validation failures, so this improves the formatting
a little.

Additionally, any validation errors that were caused by attributes in
the ACCEPTABLE_INVALID_ATTRIBUTES list are removed from the log, as
these wouldn't have prevented the activity being created.