Report state content updated
The content on the text provided to the users when they are about to change the
state of a report or they have just changed it has been updated following the
content guidance provided by the client. This gives all users a clear indication
of what will happen and what happens next when changing the state of a report.

There is a couple of test examples that needed amendment after the changes in
the content. Those changes are also included in this commit.