Display the parent activity's RODA Identifier
When entering a RODA Identifier, there's nothing to tell the user where
they are, and the data they enter here is final, so it's both easy and
very costly to make a mistake and enter the wrong thing.

As a quick improvement, we can add the parent activity's RODA identifier
to the page, if the activity has a parent. You can only enter a RODA ID
if the parent activity has one completed, so this is guaranteed to
always have a value to display.