Ingest transactions when CSV has 'transaction_type'
If the input CSV contains a `transaction_type` column, we assume that
we're ingesting transactions only, so the Transaction associated with an
Activity is modified, rather than the Activity itself.

We assume that the combination of delivery_partner_identifier and
description is unique enough to allow existing transactions to be
overwritten with new data.

When creating a transaction we attempt to associate it with the correct
report. Skipping the validation check for report means we can still
save the transaction on the occasions one doesn't exist.

Since we only know the providing organisation side of these
transactions, we also skip the validation check on
receiving_organisation_name and receiving_organisation_type so we can
save the record.