`recipient_country` list amended
When the client provided us with an up-to-date list of countries and regions to
be included in RODA, we updated the existing lists for both, and the
`country_to_region_mapping` yml file. We missed to include `West Bank and Gaza
Strip` on the list, probably due to ambiguity on the naming, as it appears as
`Palestine` in the IATI list, under the same country code.

We are now amending the `recipient_country` list to include this country, under
the name `West Bank and Gaza Strip (Palestinian territory)`. This is the name
the client wants us to use on the app. In order to be consistent on the changes,
I have modified the IATI country codelist to match the name, as we use this list
on the `activity_presenter` to show the info in the activity details.