Report the user type to analytics
The team want to use Google Analytics and understand which users are
doing what actions to help better understand how users are interacting
with the service.

As we use Google tag manager, the first step is to add the which type of
user is signed in (BEIS or non-BEIS) to the dataLayer so the tag manager
can pick it up and pass it to Google Analytics as a 'dimension'.

The principal is to create a dataLayer object before the tag manager
code is fired and add the data to that.

Although we are adding further identifiers to the tracking, I don't
believe it to be personally identifiable, I have however, made it clear
that BEIS must understand this change and any associated risk.

I originally had this as a boolean 'service_owner` but once in GA it
just would not segment the data the way the team wanted, so I ended up
using the strings in the commit.