Add special cases for Sector and Country lookups
There were a small number of ocassions where the CSV contained sectors
or countries that were not exact matches for those specified in the IATI
approved codelists.

Therefore, some additional mappings were manually made to allow ingest
to complete successfully.

NB: Some forthcoming work will ensure that the codelists are updated in
RODA so that activities can be assigned correctly.