Do not automatically strip letters from a monetary value
Previously, letters accidentally entered in the `value` fields for budgets,
transactions and planned disbursements were being stripped and the numbers
accepted.

We would prefer these entries to be rejected as an error instead of silently
stripped and passed through.

Use the `ConvertFinancialValue` to sanitize the values before they are added
to their parent entities, and show a relevant error message if any letters are
present.