Refactor organisation controller
I don't like this organisation controller doing so much work for the xml
downloads, but it is debt I think we'll have to hang on to for a little
longer.

We have to load project and third party projects to know if we should
display the 'download' button, however we don't need to map over them
making presenters as we don't show them on the view.

We might as well only show the button if the activities are publishable,
no point a user downloading an empty xml.