Show an error message if the user does not select a recipient country
Previously we were not showing an error message if the user did not select a
recipient country.

If no country was selected, we were attempting to find the associated recipient
region despite the country being blank. This was causing an error.

How, if the country is blank, we assign it to the activity, thereby forcing a
validation error to be shown. We do not try to associate the blank country value
to a region.