Refactor find programme service
Allow it to take an optional fund id and if supplied, use it to return
only the programmes who's parent is that fund.

This commit also removes the if else block in favour of a guard that
returns all the programmes early if the organisation is the service
owner. The `all` is not required as the Policy scope already returned all
of the matching programmes so is also dropped.