Users can download xml of programmes (level B)
The programme xml is split by fund (level A) so in order to show two
elements in the UI we first have to calculate which funds the
organisation is extending the programmes of.

With this information, display buttons for each fund in the UI and then
use the find programme service to load the appropriate activities,
passing in the fund id.