BEIS reference is set when uploading transactions
While testing transactions uploads, one of the devs realised that the
`providing_organisation_reference` value was not getting set on the transaction.
We know this is always BEIS, and their IATI reference is `GB-GOV-13`.

We have now amended the code to include this reference when a user uploads
transactions on RODA.

I have also changed the way the currency is set in the attributes. It was
previously set as a constant for `GBP`, but I noticed that the organisation
model gives us `default_currency` as an option, so I thought it would be more
appropriate to get the value from the model than from a default constant.