Intended Beneficiaries should be separated by a semicolon in report CSV
The Intended Beneficiaries on an activity are stored as an array. In the UI we
display the countries as "Country, Country and Country". However in the Report
CSV these need to be expressed as "Country; Country; Country".

This commit introduces an ActivityCsvPresenter which inherits from ActivityPresenter
and overrides `#intended_beneficiaries` to generate the list of countries
separated by a semicolon.