Users can edit invalid activities
Updating activities through the wizard should behave in the same way as creation
- only validating attributes relevant to the current form step

Currently the wizard behaves differently when creating and updating activities:

- When creating a brand new activity, the wizard validates only the attributes
  that appear on the current form step, which makes sense

- When updating an activity, we validate against all validation rules,
  rather than those relevant to the current step.

This was fine when activities could only be entered through the wizard,
as each step required the previous steps to all validate before moving on. Now,
we have incomplete/invalid activities - so you can be on one page of the wizard,
fix the data, and will receive an error message triggered by a validation error
caused by an attribute on a completely different form step.

This commit introduces changes so when the user needs to edit an invalid activity
they will be able to do so without getting any errors.

Co-authored-by: Leeky <leeky@leeky.org.uk>