Allow flashes to act like GOVUK error summaries
Our custom flash styles used class names that did not match those
generally passed to the flashes themselves. This led to code to
translate the flash name into a specific css class name that had the
same meaning. This commit removes that code and updates the styles to
handle the flash names.

It also updates the markup to follow the GOVUK error summary pattern.

Finally it allows us to pass a hash into a flash so we can show a error
title and list of errors, again like a GOVUK error summary.

With these changes in place we can use this new style of flash to let
the user know why a report cannot be activated, whereas previously we
could only tell them it could not be activated.