Fix bug that prevents report being submitted
Previously we were validating the Report models deadline was not in the
past at all times. This led to delivery partners not being able to
submit a report on the day of the deadline because we were setting the
deadline to 12:00 on the date of the deadline.

The fix here does two things:

1. Introduces a validation context 'edit' that is used on the edit
   report controller so we can ensure the deadline is validated
2. At all other times the deadline no longer matters (we do not enforce
   the deadline) it is for information only.

Now the Report objects are valid regardless of the deadline date (or
time), unless they are editing those details, the users can submit their
report successfully.

This commit also sends a validation context when changing the state of a
report. The context name is passed from the `policy_action` so:

- activate
- submit
- review
- request_changes
- approve

This lets us run specific validations on reports as the state changes.

In this commit we validate the presence of a report description when
the state changes to active. This keeps the behaviour as it was
previously, although this may not be the actual behaviour we want the
user to experience, but we can fix that in later work as the priority
here is to fix the bug.

There are two other changes.

1. A change to the error message content
2. The save! is not required as we just called valid?