Planned disbursement type is always original
When creating a new planned disbursement (forecast) the type is always
'original'.

Later work will introduce the concept of 'revising' a forecast.

As a fist step to this the create planned disbursement service sets the
value of planned_disbursement_type and we no longer show that section of
the form in the application.

For now we will continue to store the value as 1 or 2, later we may
change this and potentially not store this value at all as we can
calculate it by other means at the point we publish to IATI.