Set currency on a new planned disbursement
Different currency is not something will can support for some time yet.
The currency should always be set to pounds sterling, stored as GBP.

As we set the default currency on the organisation already and we have
to have an organisation to add anything, we can use the Organisations
it to set the value of currency when creating a new planned
disbursement rather than introducing another place where the default
currency is set.