Planned disbursement has no receiving organisation
The decision has been made to not collect a receiving organisation for
planned disbursements as this would introduce a great deal of complexity
not just for the application but for a delivery partner user having to
forecast spend for every recipient instead of the total spend for an
activity per financial quarter.

As a first step we remove the validation that requires a receiving
organisation name and type for a planned disbursement. We also remove
the fields on the form.

We have not removed the columns at this point as we have data in
production with the value set so there seems no reason to loose that at
the moment.