Stop modifying the Bullet whitelist during tests
A few of our tests have this configuration that seems to designed to
make Bullet temporarily ignore certain optimisations:

    before :all do
      Bullet.add_whitelist(type: :unused_eager_loading, class_name: "Activity", association: :parent)
    end

    after :all do
      Bullet.delete_whitelist(type: :unused_eager_loading, class_name: "Activity", association: :parent)
    end

This particular setup makes Bullet ignore unused eager loads of the
`Activity#parent` association. The problem is that, if this assoication
was already ignored before these tests ran, then the call to
`delete_whitelist` will cause it _not_ to be ignored in any tests that
run later.

This is because of how Bullet stores its whitelist. It indexes
associations by optimisation type and class name, so if we have this in
the application init code:

    Bullet.add_whitelist(type: :unused_eager_loading, class_name: "Activity", association: :parent)
    Bullet.add_whitelist(type: :unused_eager_loading, class_name: "Activity", association: :organisation)

Then `Bullet.whitelist` will contain:

    {
      :unused_eager_loading => {
        "Activity" => [:parent, :organisation]
      }
    }

The association names are stored in an Array, not a Set, so it's
possible for them to appear twice. When we run the `before` block above,
`:parent` is added again to this list:

    {
      :unused_eager_loading => {
        "Activity" => [:parent, :organisation, :parent]
      }
    }

When we run the `after` block, `delete_whitelist` removes _all_
occurrences of the given association from the whitelist, so the state
ends up containing:

    {
      :unused_eager_loading => {
        "Activity" => [:organisation]
      }
    }

The `:parent` association is no longer ignored and so all subsequent
tests may trigger warnings about it. Because RSpec randomises test
order, this means we'll see inconsistent results with Bullet causing
different tests to fail each time.

It doesn't look like Bullet provides a safe API for this use case of
temporarily ignoring a certain warning type. `delete_whitelist` itself
is undocumented so it's unclear we should even be using it, or what it's
intended to do, but it doesn't provide the inverse of `add_whitelist` as
it removes _all_ occurrences of the given association from the list.
`add_whitelist` doesn't provide any indication that the association
we've just added was already in the whitelist, and therefore we
shouldn't remove it after the test.

The safest thing to do is probably to only use `Bullet.add_whitelist` in
the application config, and not in individual tests.