Rake task for invalid activities
The original plan to flag invalid activities was to introduce a temporary change
in the UI to inform the user what activities were invalid. Once the activities
had been amended, we could undo the changes on the code and leave the app as it
was previously.

After discussion with the wider dev team, we agreed that a better approach for
this feature would be creating a rake task that we could run in production in
order to get a list of invalid activities. This list will contain the organisation
name, the title of the activity, the level, and the URL of the activity. We can
then supply this list to the client, so they will be able to amend those
activities.

The reason for this approach is because it doesn't require to make changes to the
codebase that later will need to be reverted, as this takes time. It is straight
forward enough that a rake task could give us the same outcome, with less risks
and being less time-consuming.

This rake task stores a csv file with the invalid activities in the tmp folder.
We have tested this on local environments and also the spec asserts we get what
we need.