Options for GDI form step have been modified
This follows a request from the client. They needed to include one option
for GDI for activities that benefit either India or China (or both) but that
were created before the GDI policy existed.

Likewise, the option `No` is now presented as `GDI not applicable`, which will
be the option users can select for activities in all other countries.

Specs and factories have also been modified to reflect these changes.