Modification on failing spec after changing GDI
After introducing the new values for GDI form step one of the ingest specs failed
This commit introduces changes to the spec to reflect the new state of GDI.

Since the ingest work has been done by another developer from the team, I decided
to put this file on a separate commit in case they want to make some changes to
their work around GDI. The main possible issue here is that one of the GDI
options now is `GDI not applicable`, which is very similar to the `not applicable`
scenario reflected on this spec. However, these two `not applicable`s mean
different things now. This will need to be discussed with the relevant developer.