Documentation for rake task
Following a suggestion from another developer, we are now including a brief
guide to the rake task we created. This will give some context to any potential
new developers joining the team or anyone looking at the code. I believe it is
important to keep documentation clear and updated for other people. The file
introduced by this commit should tick those boxes.