Only allow one original planned disbursement
There should only ever be one original planned disbursement for a given
activity, financial quarter and year. This commit enforces that with a
validation on planned disbursement type that only apples when creating a
new planned disbursement.

This also meant updating a number of specs that are creating multiple
planned disbursements in the same financial quarter.

As there is no field in the UI to link an error to we add the error to
the base and use the GOVUK form builders link_base_errors_to option to
make the link the financial quarter, however there seems to be a bug
with this that the id for the link is not generated correctly, it is
missing the `--error` part- I don't feel that this should block this
work however.