`oda_eligibility` is now an integer-enum
Following a request from the client to add a new option to ODA eligibility,
it would not work as a boolean anymore since now we have 3 possible options for
this field. I have converted this column to an integer, and made it an enum so
we will have helper methods that could be handy.