New `oda_eligibility` option added
Before this change, `oda_eligibility` was a boolean with a value of `true` for
eligible activities, and `false` for no longer eligible activities. The client
wanted an extra value added (No - was never eligible) to report activities that
never complied with the standards of eligibility.

Having this changed meant that we needed to change the column type in the database
(previously a boolean), which is reflected on the previous commit. We now need
to include the new option on the UI, and adapt the code to reflect the way we
store these values.

For a similar reason, I have also made some changes on the specs.