Fix Auth0 create_user call
We recently updated the Auth0 gem and I missed a breaking change.

The signature of the `create_user` method has changed; the connection
name, which is just a string that matches the default database
connection in Auth0, comes first and is required and the name of the
user is passed in the options as `name`.

Without this change the application cannot create new users in Auth0.
The application handles the error gracefully as the API simply returns
an error code that we display to the user.

We've update the stubbed request to require the connection now to catch
this.