Adding RODA IDs to invalid activities list
After supplying the client with a list of invalid activities, we were asked to
add the RODA ID of each activity, so the users could group them up and make it
easier for them to work on them.

This commit adds this RODA ID for each activity on the csv list.