Show Free Standing Technical Operation status
This adds a new row to an activity's details page to show the FSTC status of the
activity. Because the column is boolean, I've also tweaked the activity presenter to
show "Edit" if the column is set to false. Before, setting the column to `false`
meant that `present?` returned `false`. Casting the column as a string means that
if the column has been completed and is `false`, the column shows up as needing
to be edited.