Policy markers created as enum
Policy markers will be a new field in the activity form. Users will be able to
select an option for 8 different markers stated by IATI. The options are the
same for all the markers, so it made sense to make these markers as `enum` and
be able to have the helper methods that come with them.

Since Rails does not allow to share enum methods in multiple database columns,
I am using prefixes to work around this (see changes on the activity model).