Policy markers added to the activity form UI
This commit introduces the 8 policy markers in the UI for the activity form.
They are different entities but we think that grouping them together on the same
page makes more sense and it gives the user a better experience. They are select
boxes to avoid having a large selection of radio buttons, and the default value
that the customer asked us for is already pre-selected on the box, so in case
the user does not need to change any of these values, they just need to click on
`Continue` to save the 8 values at the same time.

Please note that according to validation rules on the previous commit (activity
model), policy markers are only used for levels C and D, so we are only displaying
them on those activities.