Policy markers get reported to IATI
The 8 new policy markers get reported to IATI so we have now included them on the
activity XML file that RODA creates for each activity. We do not report those
markers that are under the option `not assessed`, as this is a BEIS custom
option for internal use only, and not something IATI recognises. Other than that,
we report the values for all of them.

I have checked some activities XMLs with different combinations of values for
policy markers and they pass the IATI validator.