Add 'sdgs_apply' boolean attribute
I originally tried doing this with a virtual attribute to avoid needing
a dedicated database attribute to store whether the Sustainable
Development Goals were relevant or not.

My theory was, this state was only useful when showing the form, because
`sdg_1` was only really required when the radio button option "the
sustainable development goals to this activity" was selected.

In the end, I couldn't get the wizard to behave itself, so opted for the
simpler solution of having a boolean `sdgs_apply` attribute on
`Activity`.