Start building out the Activity import machinery
This adds a couple of classes that start to build out the code that will eventually
backfill missing activity data. So far it accepts an array of hashes (which will
eventually come from a CSV) to update an existing Activity (which is fetched by
its RODA ID). If a field is blank, it is ignored. If any activities in the array of attribute
hashes are missing, then the entire operation fails and the main classes error
array is populated.