Simplify the activity updating
This adds a new method to ImportFromCsv that initializes the ActivityUpdater,
calls `.update`  on it and returns the class instance. If there are any errors, the
importer loops throuhg them and adds the error to the base class instance.

If there is no RODA ID, an error is raised.

This sets the groundwork for adding a new class that will create an Activity in the
next commit.