Allow multiple planned disbursements for the same activity and quarter
This reverts commit 12858190ef21b7c3680e0f41c520f7410faf30eb, which was
added to ensure that while forecasts cannot be revised, we only allow
one forecast per activity per quarter to be created.

To allow forecasts to be revised, we need to allow creation of multiple
forecasts per activity per quarter, and this validation rule prevents
that.

Rather than treating this as a problem of validating user input, we're
instead heading towards a model where the app always records forecasts
in the appropriate way to capture the revision history, rather than
making end users figure out how to do that.