Allow importer to create Activities
The importer can now create activities which:

a) Do not have a RODA ID specified and:
b) Have a RODA Fragment and RODA Parent ID

If the above is true, the row is passed to the ActivityCreator class, which initializes
a new activity, assigns the attributes, saves it, and gathers up any errors.

As we can only validate against a limited subset of contexts, there's a constant
that has the contexts we can validate against with the fields we have. Eventually
this will go away.

I've also added a new converter that fetches the parent Activity by its RODA ID,
and returns the UUID from the database. If the activity cannot be found, an error
is added to the errors hash.