Improve ambiguity detection
This works on the ambiguity detection when trying to sort out whether to import
or update an Activity. We first check for the presence of the RODA ID.

If the RODA ID is blank (i.e. we are creating an activity), we check if a fragment and
parent ID is present before  attempting to create an Activity, and raise an error if
not.

If the RODA ID is present (i.e. we are updating an existing activity), we check for
the presence of a fragment or parent ID,  and raise an error if either is true, as this
may overwrite existing fragments or parent IDs.