Revise planned disbursements for non-versioned activities
Here we begin the implementation of PlannedDisbursementHistory, a
service that represents the forecast for a particular activity and
quarter, and stores revisions to the value of that forecast. For
example, you can set the forecast spend for an activity in Q3 2020 like
so:

    history = PlannedDisbursementHistory.new(activity, 3, 2020)
    historical.set_value(25_000)

The job of the `set_value` method is to store this new value so that it
becomes the _current_ value for this forecast, and any previous values
are retained as necessary.

We will model history by pinning data to reports; retrieving history
means being able to see the values contained in any prior report. At
level A/B however, we don't use reports. BEIS can make arbitrary edits
at this layer and we don't do reporting cycles at these levels.

Therefore, the model at level A/B is quite simple: we only store an
"original" value and a "revised" one. The first time the value is set,
that becomes the "original" value. The second time the value is set,
that is added as a "revised" value. All subsequent writes update the
value of the "revised" record.

So in effect we only store two values for forecasts at these levels: the
original value and the latest revision.