Keep track of activities that have been changed
This adds two arrays that keep track of created and updated activities, so we can
report back what the future Rake task has done. When we successfully update or
create, we append to the relevant array. If there is are any errors, we reset the
arrays before raising a rollback error.