Link planned disbursements to a report
When a planned disbursement record is created, it may need to be linked
to a report. Data at level A/B does not go through the reporting cycle
and is "unversioned". Data at level C/D does go through the reporting
cycle, and is versioned by pinning data values to the reports in which
they were created.

Therefore, if the planned disbursement relates to activity belonging to
the service owner (i.e. a level A/B activity owned by BEIS), we don't
link it to a report. If it's owned by a delivery partner, then we link
it to the current open report for that partner.

We could also have implemented `record_history?` by checking the level
the activity belongs to, rather than checking who owns it. It should
always be the case that level A/B activities belong to BEIS, and level
C/D ones do not belong to BEIS, so these should give the same answer.