Link revised planned disbursements to the current report
At level C/D, we want to record the history of changes made to a
forecast, where each forecast relates to an activity and financial
quarter. "History" in this case means that we want to be able to
retrieve the values that were signed off in any past report. So, rather
than storing every single value saved for a forecast, we record one per
quarter.

The implementation at level A/B is "unversioned" and stores only the
original value of a forecast, and the most recent revision. The
"revision" record is created by cloning the original and changing the
`value` attribute. Forecasts are this level aren't linked to reports.

At level C/D, we want any revisions we create to be linked to the
current active report for the delivery partner, so we have a record of
the value they entered in that report. As new reports are opened, new
forecast values will be linked to them, forming a chain of history.