Build a historical log of planned disbursements
A "forecast" or "planned disbursement" is an assertion that we plan to
spend money on a certain activity in a certain quarter. That assertion
can change over time -- the amount we plan to spend may change. At level
C/D, we want to capture that change history, so we can see what the
forecast for any activity and quarter was in any previous report.

We do this by storing a series or log of values for a forecast. A single
logical forecast is identified by the activity and quarter it relates
to; in the database this is covered by the columns (parent_activity_id,
financial_quarter, financial_year). We store the history of the `value`
of a forecast by storing a set of records with the same
(parent_activity_id, financial_quarter, financial_year), where each is
linked to a different report.

As reports form a sequence in time, we can retrieve the log of a
forecast's value by sorting all the records for the same activity and
quarter, based on the report each belongs to. Reports are ordered by the
financial quarter they relate to.

As before, storing a value for a forecast is done like so:

    history = PlannedDisbursementHistory.new(activity, 3, 2020)
    history.set_value(25_000)

At level A/B, the logic of `set_value` remains the same: it creates an
"original" record if none exists yet, and creates or updates a "revised"
record otherwise.

At level C/D, we want to associate the new value with the current
report. If there's a record for this forecast in the current editable
report, then we modify it in place. Otherwise, we create a new record
containing the value.

This has the effect that any records linked to older reports are
retained any not overwritten -- we only create or update records in the
current report, and once that report is approved, its forecasts are no
longer editable.

At level C/D we still have the logic that if no records for this
forecast exist, then an "original" record is created, otherwise a
"revised" record is created. The difference between this and the A/B
logic is that "original" entries can still be modified, if they belong
to the current report.