Introduce Report.for_activity scope
A pattern I've seen across the codebase involves looking up the reports
relating to an activity, by finding the report whose fund and
organiation match those of the activity. Here we introduce a scope that
encapsulates that pattern and refactors all existing code to use it.

I've also refactored more code towards using the `editable` scope and
other scopes based on the `state` enum, to encapsulate that logic and
avoid loads of the codebase knowing about how the `state` column works.