Retrieve the forecasts for an activity
The PlannedDisbursementOverview service takes an Activity, and returns
a set of PlannedDisbursement records representing the latest values of
the forecasts across all quarters for that activity.

    overview = PlannedDisbursementOverview.new(activity)
    disbursements = overview.latest_values

It can also retrieve the PlannedDisbursement objects that represent the
forecast values as of some prior report:

    report = Report.find(report_id)
    disbursements = overview.values_at_report(report)

It does this by looking up the planned disbursements for the activity,
and for each quarter is selects the mos recent record, based on ordering
by the report each one is linked to. It uses the `DISTINCT ON` feature
in Postgres to perform a query like this:

    SELECT DISTINCT ON (pd.financial_year, pd.financial_quarter) pd.*
    FROM planned_disbursements AS pd
    INNER JOIN reports AS r ON r.id = pd.report_id
    WHERE pd.parent_activity_id = ?
    ORDER BY
        pd.financial_year ASC,
        pd.financial_quarter ASC,
        r.financial_year DESC,
        r.financial_quarter DESC

The first two `ORDER BY` terms mean the forecasts are returned in
ascending order of the quarter they related to, and the latter two
clauses mean that `DISTINCT ON` will select the value in the most-recent
report for each forecast.

Note that the PlannedDisbursement objects returned will not all belong
to the same report; it returns the most recent PlannedDisbursement
record for each forecast, and some of those may come from older reports.
We don't store values for all forecasts in every report, we only store
new records when a forecast is changed.

Getting the forecasts that were current as of a certain report means
adding a `WHERE` to the above that selects for reports sorting earlier
or equal to the given report.

    SELECT DISTINCT ON (pd.financial_year, pd.financial_quarter) pd.*
    FROM planned_disbursements AS pd
    INNER JOIN reports AS r ON r.id = pd.report_id
    WHERE pd.parent_activity_id = ?
    AND (
      r.financial_year < ?
      OR (r.financial_year = ? AND r.financial_quarter <= ?)
    )
    ORDER BY
        pd.financial_year ASC,
        ...

At level A/B where this versioning is not applicable, we instead order
the records by planned_disbursement_type, as "original" is stored as 1
and "revised" as 2, so "revised" sorts after "original".

    SELECT DISTINCT ON (pd.financial_year, pd.financial_quarter) pd.*
    FROM planned_disbursements AS pd
    WHERE pd.parent_activity_id = ?
    ORDER BY
        pd.financial_year ASC,
        pd.financial_quarter ASC,
        pd.planned_disbursement_type DESC